fix gk_sort problem in green_linsys_shift_im
When a k point not present in the ground-state calculation is used there is no
guarantee that |k + G| < Ecut is fulfilled by less than the maximum number of
G's that fulfilled this condition for the original k-point mesh. As a
consequence it happened that the code crashed with an out-of-bounds exception.

The new wrapper routine gk_sort_safe avoids this by temporarily increasing the
array boundary by 10% to call gk_sort and resetting everything to the original
values afterwards. G's that are outside of the original maximum are discarded.

Possible improvements:
- Determine actually needed npwx_gw at the beginning of the calculation so that
  we don't need to reallocate the arrays.
- Possibly other routines need the gk_sort_safe wrapper for the same reason.

git-svn-id: https://rothery3.materials.ox.ac.uk/svn/sgw/branches/sgw-5.0@803 21dd5d8e-1f79-4369-87ad-170e0d29e828