introduce 8-byte integer for record lengths
For files with very long records a 4-byte integer might not be
sufficient so that the 8-byte integer is used. Previously this was
done using INTEGER*8 which is not conform with the FORTRAN standard.
Hence, we introduce a module kinds_gw (analogous to kinds of QE)
that contains a i8b type that can be used to define long integers.

Replace INTEGER*8 with INTEGER(i8b) in all files.

git-svn-id: https://rothery3.materials.ox.ac.uk/svn/sgw/branches/sgw-5.0@819 21dd5d8e-1f79-4369-87ad-170e0d29e828