encapsulate pp file in type
To allow for easier modification of the metadata of the different PP files,
we encapsulate everything in a single type. Currently, we add the number of
bands and k-points as well as the unit once the file is opened.
Remove the check if the filename is empty from gwq_readin. It will be
introduced in the pp_output_module later.

git-svn-id: https://rothery3.materials.ox.ac.uk/svn/sgw/branches/sgw-5.0@837 21dd5d8e-1f79-4369-87ad-170e0d29e828