fix of nbnd_occ allocation
Due to a recent change in QE, nbnd_occ must be allocated at the beginning of
the calculation and is not fixed size anymore.

We replace a major part of the gwq_setup routine by calls to functions inside
the LR_Modules. This also takes care of the allocation of the nbnd_occ array.
To ensure that the calculation can run at multiple points, we need to deallocate
the nbnd_occ array in deallocate_gwq.

git-svn-id: https://rothery3.materials.ox.ac.uk/svn/sgw/branches/sgw-5.0@873 21dd5d8e-1f79-4369-87ad-170e0d29e828