add routine to expand igk by one element
To replace h_psiq with h_psi and s_psi, we need to eliminate the igkq dependence
of the code. Hence, we need to store those elements as one column of the igk_k
array. To make this possible, we write a function to expand the igk_k array by
one element.