unitfy para_img and para_pool in parallel_task
para_img and para_pool conceptionally did the same thing: they split a number
of tasks equally across the process grid of their respective communicator
(images or pools) assigning excess tasks to the first processes if necessary

In the new file parallel.f90 we provide a routine parallel_task that takes the
communicator as an additional argument so that the same routine can be used for
both tasks. We change the assignment of the excess tasks to the last one,
because the head node is often used to do extra work (printing, io, ...) so it
is better if he finishes earlier. In addition, we return the number of tasks
done at the other processes so that we can replace MPI_ALLREDUCE with
MPI_ALLGATHERV later.

We adjust all files to make use of the new routines.

This is the first file that is tested with the pFUnit testcase. So that we
include not only the testcase but also the necessary Makefile to build all
testcases.