replace the igk and g2kin initialization
igk is now stored in a module so that we just have to access the appropriate
igk_k at k and k + q. Note that for now, we still initialize igk and igkq,
because they are used by other parts of the code.

g2kin is now initialized by the function g2_kin, so that we don't need to
calculate the kinetic energy in solve_linter anymore