introduce the BiCGstab solver
We use the BiCGstab solver for all cases in solve_linter. Currently, we make
no use of the fact that this is an iterative scheme which could be speed up
significantly by reusing data from a previous iteration.

For the first iteration, the frequencies are degenerate, so that we use a
multishift solver. From the second iteration onward we solve each frequency
on its own.

So far we didn't test the routine with multiple frequencies yet, but a large
part of the code is already written to take into account the possibility to
do multiple frequencies.

The calling coulomb routines are modified to pass the frequency instead of the
frequency index. Furthermore, we replaced some statements that are not allowed
in Fortran 2003 with standard-conform alternatives.