enable the call with multiple frequencies
When the linear solver is called with an array containing several
frequencies, we solve for all of them at once. This allows us to
use the multishift solver in the first iteration.

Adjust some array sizes to include the frequency dimension and/or the
frequency index in the call of subroutines.