replace direct solver with solve_linter
The new implementation of the solver results in the direct solver if the number
of iterations is set to 1. So we replace the calls to solve_direct with calls
to solve_linter with the number of iterations set to 1. In solve_linter, we
check if only 1 iteration was specified and switch to the direct solver in this
case. This means in particular, no mixing and that the result is -dV instead of
+dV.