implement calculation of Sigma correlation
we reimplement a large part of the functionality of sigcim.f90

For a given set of frequencies, we evaluate the Green's function and Fourier
transform it to real space. Then we construct W at the necessary frequencies
to convolute G and W to obtain Sigma.

The differences of the new implementation are:
- the new BiCGstab solver is used for the Green's function
- the Fourier transform of G is only evaluated once instead of inside the loop
- a better documentation and no dependence on external modules

We have already included the code to treat real frequencies though it is
not tested yet and hence commented out.