fixed a few small bugs in exchange
The documentation was rewritten, because the old one lead to a small but significant
bug in the implementation. The new theory does not explicitly rely on the Fourier
transformation.

The eigenvector is extracted at G - G'' and G' - G''. The latter was swapped which
lead to incorrect exchange elements.

The loop in the convolution routine should only run over the elements of G''.

Directly accumulate Sigma, i.e. include the sign in the summation.
Rescale the value by the volume at the very end.