begin creation of unit test for linear solver
The file produced by the debug routine is read in a new utility module. We add
a new test suite that currently only reads the file with this new module.

The rest of the changes are updates to the Makefile to facilitate the build of
the utility routines and the linking to it.