add option to track second array to Gram-Schmidt
If two sets of vectors v_i and b_i are related by a linear operator A v_i = b_i
and we use b_i to construct an orthonormal basis B_i, we change v_i accordingly
to V_i such that A V_i = B_i.

A unit test for this transformation is added.