restore previous behaviour for coulpade
We needed to restore some previous behavior, because the parallelization of the
G vectors does not guarantee that all symmetry-equivalent points end up on the
same process. Hence, we do some extra work for W (using all G vectors) and only
symmetry reduce it when evaluating the coefficients. The idea is to remove
gmapsym from the sigma_prod routine.

NOTE: image parallelization still broken