fix the image parallelization
We reorder the resulting self-energy and collect it by a mpi_sum on
the root process. This is not the best implementation as we communicate
a lot of zeros. A better implementation would use mpi_gather, but that
would require a rearrangment of the array.
In the long run a parallel I/O would be the best choice to avoid the
communication alltogether.