construct wrapper for different solvers
We want to be able to dynamically change the linear solver if the selected
solver did not converge in the specified number of iterations. The idea is that
the code detects if the linear system was solved successfully and only stops
in this case. Otherwise it will try again with a different linear solver.

In this commit, we write the outline of the wrapper module without the actual
implementation of the calls to the wrapped routines.