bugfix for pade_robust
For functions where the output degree was different then the input one
the pade_robust function did not work, due to an error in the allocation
of the matrices for the Toeplitz matrix. We changed back to the original
implementation in chebfun to fix this.

Add test case for Pade expansion of cosine function and compare to MATLAB
results.