new concept for real frequencies
Conceptionally, one can image the frequency convolution on the real axis in two ways
i) We approximate G and W on the real axis by moving slightly above the real axis
   so that we can actually calculate the quantities.
ii) We actually are slightly above the real axis and choose the frequency of W
    such that the self energy is on the real axis.
When refactoring the code to the new version, we had moved from method (i) to method
(ii) but this requires some care for the frequency of the Green's function. The non-
analytic part of the Green's function corresponds to taking the frequency at the
opposite point of the real axis.

Hence, we change the frequency setup to be above/below the real axis depending on the
sign of the frequency of the Green's function. Whether W is evaluated for positive or
negative imaginary part follows directly from this. This change also includes the
nonanalytic part, so that a separat treatment of this term is no longer necessary.