add new test case for real frequency integration
Use one of two Pade approximations to go from an initial frequency grid to
the real axis, then generate the Green's function there to convolute it.

The test case performs this for diamond with conventional and robust Pade
approximation. The second case is restarted from the W converged for the
first one.