improve stability of BiCGstab solver
For shifted system a few variables got very large so that we replaced them
with their inverse which will go to 0 instead. This is numerically more
stable, because the product of two very large number can get out of bounds
whereas the product of two small numbers is rounded to 0.