extend the Coulomb matrix using the symmetry
use W(w) = W*(-w) to expand the Coulomb matrix array to twice the size to
improve the accuracy of the Pade approximation

Because freq_symm is the new default, we need to modify all tests to reproduce
the old results by setting freq_symm to false. Note, that in the future the
new symmetry should be introduced for some of the tests.

Currently, the new symmetry is not working yet, because coulpade is not
adjusted to the bigger Coulomb array yet.