move frequency symmetry to coulpade
After convoluting eps^-1 and V we extend the reduced frequency mesh to the
full one. At the same time, we extend the frequencies to the full mesh, which
is something, we were not able to do in sigma. As a consequence the array
dimension of the symmetrization shrinks to a 1d array.