merge functionality of routines in freq_symm
The extension of the frequency mesh was handled in two different routines
that did the same thing, which could easily go out of sync. Hence, we merged
both routines into a single one that can deal with both cases which improves
the stability and makes it easier to switch to a more sophisticated method
in the future.
Adjust the calling routines in construct_w and coulpade.