bugfix: write whole Coulomb matrix before reading
On some systems the Fortran WRITE command only puts the data in the buffer
(even with direct access) and the file is not complete until it is closed.
The other processes the tried to read an element that was not written yet.

The solution is to close the file and open it for reading at the appropriate
place afterwards.