make no frequency symmetrization default
The frequency symmetrization may not be fully stable at the moment. It should
work for the imaginary frequency integration, where all points should have no
imaginary part. However, for points away from the imaginary axis, an additional
transpose might be necessary, which is not implemented at the moment.

Hence, we switch back to the old behavior which should still work if one
manually adds the additional frequency points.