deactivate alpha_pv for metals
There should be an extra contribution for metals because the projector
would become frequency dependent. As this extra term is not implemented
yet, we set alpha_pv to a very small value so that it doesn't distort
the results. We need to check how this performs for a real metal.