restrict construction of symmetry map
We only need the symmetry map of the G vectors up to the actual value
used in the calculation. The construction for the whole mesh is much
more expensive, so that it became noticeable for some systems. Hence,
we construct only the elements that we actually use to save some time
and increase the parallel efficiency.