fix list in input generator
fix1: YAML specifies lists as - instead of *
fix2: Markdown requires empty line before list (at least in FORD)
fix3: Tiddlywiki requires empty line before list and * as list symbol

The first fix is done in the YAML file gw_input.yml. The other two
amount to detecting a list with the input generator and adding a new
line if a list is recognized.

We also add a new line where there is a line break in the YAML file
for the source code which avoids overly long comments.