adjust thresholds so that the test suite works
Some of the extracted quantities were so sensitive to small changes in the
calculation that they did not reproduce the same results on different
machines. One example would be the real and imaginary part of the self energy.
Both exhibit zeros so that the relative error could become very large or
the pole shifted slightly which lead to large absolute errors. Visually the
curves looked the same, but comparing at the same frequency did not work.
Hence, we reduce the testing to the qp eigenvalues and the spectral function
that are more stable.