replace logical with integer for screening model
Instead of a nested if tree, we use a select case construct by introducing an
integer to represent each of the possible choices. The possible values are
stored in a new module in the file analytic.f90 and all routines using the
flags are changed to make use of the new integer logic instead.

In addition, we introduce a constant for the new AAA approximation and document
it in the input file.